Massive code cleanup

+++ src/main/kotlin/dev/luna5ama/trollhack/module/modules/misc/AutoObsidian.kt
@@ -604,9 +604,9 @@ internal object AutoObsidian : Module(
                                 pos,
                                 side,
                                 EnumHand.MAIN_HAND,
-                                hitVecOffset.x.toFloat(),
-                                hitVecOffset.y.toFloat(),
-                                hitVecOffset.z.toFloat()
+                                hitVecOffset.x,
+                                hitVecOffset.y,
+                                hitVecOffset.z
                             )
                         )
                         player.swingArm(EnumHand.MAIN_HAND)
